modified folder structure: moved modal component and corresponding css file inside modal folder and fixed isOpen prop type inside ImageUploadModal component
